Fix webpack for dev
