+++ admin-dev/themes/new-theme/.webpack/prod.js
@@ -1,9 +1,8 @@
-const fs = require('fs');
 const common = require('./common.js');
-const MiniCssExtractPlugin = require('mini-css-extract-plugin');
 const UglifyJsPlugin = require('uglifyjs-webpack-plugin');
-const OptimizeCSSAssetsPlugin = require('optimize-css-assets-webpack-plugin');
 const keepLicense = require('uglify-save-license');
+const OptimizeCSSAssetsPlugin = require('optimize-css-assets-webpack-plugin');
+const cssExtractedFileName = 'theme';
 
 /**
  * Returns the production webpack config,
@@ -12,8 +11,6 @@ const keepLicense = require('uglify-save-license');
  * @param {Boolean} analyze If true, bundle analyze plugin will launch
  */
 function prodConfig(analyze) {
-  const cssExtractedFileName = 'theme';
-
   let prod = Object.assign(
     common,
     {
@@ -32,6 +29,7 @@ function prodConfig(analyze) {
           }
         },
         minimizer: [
+          new OptimizeCSSAssetsPlugin(),
           new UglifyJsPlugin({
             sourceMap: true,
             uglifyOptions: {
@@ -43,57 +41,16 @@ function prodConfig(analyze) {
               }
             },
           }),
-          new OptimizeCSSAssetsPlugin()
         ]
       },
     }
   );
-
-  prod.module.rules.push({
-    test:/\.(s*)css$/,
-    use: [
-      MiniCssExtractPlugin.loader,
-      'css-loader',
-      'postcss-loader',
-      'sass-loader'
-    ]
-  });
-
-  prod.plugins.push(new MiniCssExtractPlugin({
-    filename: '[name].css'
-  }));
 
   if (analyze) {
     const BundleAnalyzerPlugin = require('webpack-bundle-analyzer').BundleAnalyzerPlugin;
 
     prod.plugins.push(new BundleAnalyzerPlugin());
   }
-
-  prod.plugins.push({
-    apply: (compiler) => {
-      /**
-       * When using mini-css-extract-plugin and merging all chunks to one file (see optimization configuration),
-       * a [cssExtractedFileName].bundle.js is created. This file is required for the js entry point to be executed.
-       * see: https://github.com/webpack-contrib/mini-css-extract-plugin/issues/147
-       * This hook merges the [cssExtractedFileName].bundle.js into the main.bundle.js file, so we avoid
-       * to include the [cssExtractedFileName].bundle.js into the html
-       */
-      compiler.hooks.afterEmit.tap('AfterEmitTest', (compilation) => {
-        let mainBundle = fs.createWriteStream('./public/main.bundle.js', {flags: 'a'});
-        let themeBundle = fs.createReadStream('./public/'+ cssExtractedFileName +'.bundle.js');
-
-        mainBundle.on('pipe', function() {
-          console.log('prestashop-post-operation: Merging bundle.main.js and '+ cssExtractedFileName +'.bundle.js');
-        });
-
-        mainBundle.on('close', function() {
-          console.log('prestashop-post-operation: Merging done.');
-        });
-
-        themeBundle.pipe(mainBundle);
-      });
-    }
-  });
 
   return prod;
 }
